+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Added in Unreleased
 
 - `NestTupleRef` and `NestTupleMut` traits for converting flat tuples into nested tuples of references and mutable references, respectively.
+- `NestedTupleFlattenOption` trait for flattening nested tuples of double options into nested tuples of options.
 
 ## [0.1.4] - 2025-12-15